Try to improve lifecycle
CefInitialize() / CefShutdown() cannot be called multiple times during
process lifetime making the "cefsimple" example architecture insuitable for use
in plugins without modifications.

A workaround could consist in moving the CEF main loop to a separate
subprocess, then have the DPF plugin communicate to it via IPC. This
would avoid the above issue and also keep all plugin instances
completely isolated, including the potential issue where two different
plugins try to load CEF within the same host, that would not work.

Idea comes with its own challenges, the plugin will need to reparent a
window from a different process. That is known to work for X11 and
possibly Windows but not sure on macOS. A separate window would still be
fine so it is not really a terrible drawback compared to the benefits.